Add some changes to make connection recovery easier for senders (#42)
As part of an investigation into an issue in Event Hubs I found a couple of cases that were problematic when attempting to do recovery.

First, if an error happens (and the sender is subsequently 'detached') that same error will get repeatedly returned by sender.Send(). The error itself is actually a throttling error (and was relevant the first time it was returned) but after that we really should be indicating the link will _never_ work again because it has been detached (ErrLinkDetached). 

The second is just exposing the link ID so it's simpler, from the outside, to tell one sender apart from another. As part of a separate change I want to make in Event Hubs I need to be able to see if the sender that I'm about to close/remove is actually the same one I started with (if it's not then I don't have to recover since the 'bad' link has already been cycled out).